v-1.0/ai-ml/assistant/requirements.txt
Flask
pymongo
python-dotenv
pydantic
pytz
pypdf
langchain
chromadb # Vector storage
pytest
boto3
langchain_community